Substantial changes for new message names and signup interface, but also to move the location of the offset point to the upper left, opposite the handle.  This complicates some of the boundary calculations, but makes box resizing and moving much more straightforward.
Heavier use of the tvDraw::Convert for coordinate conversions.  No
boxflag=2 required any longer, since drawing a new box is exactly
equivalent to resizing the box.  The box now can be expanded or initally
drawn only down and to the right.

Added the ON_MOTION keyword to allow box events to be sent when the size
or coordinates of the box change at all, not just on button release.  Be
careful with this one, as it generates a lot of events.

Arrow events are now requested from a Key object (if present, ala
self.oDraw->GetMsgObjs('tvKey')), which step the box by one data pixel
in the indicated direction.  Probably it should have a way to disable,
in case someone else wants exclusive control of the arrow keys.