The vestigial TVDRAW_EXCLUSIVE messages were completely eliminated, in favor of all object selection driven directly by tvSwitcher, which holds the buttons, and responds to the key events, and is therefore a much more natural place for this logic to be.
In addition, the tvSwitcher object now takes a MsgList structure with
an "Exclusive" field (in addition to object and activation key) which
indicates how that object should be treated when a key or button press
event occurs.  Exclusive=1b is the old behavior, and exclusive=0b
describes objects which can just be "toggled" on and off,
independently of any others.  As always, the methods "On" and "Off"
are required for objects wishing to receive the services of tvSwitch.